requirements.txt
aiogram==3.27.0
maxapi==0.9.18
asyncpg==0.31.0
SQLAlchemy==2.0.49
alembic==1.18.4
pydantic-settings==2.13.1
python-dotenv==1.2.2
aiohttp-socks==0.11.0
cryptography==46.0.7
prometheus-client==0.25.0
tenacity==9.1.4
aiofiles==25.1.0
